Added the SkillEntity refactorings

+++ swords-skill/src/main/java/com/morethanheroic/swords/skill/service/handler/SkillHandler.java
@@ -9,7 +9,7 @@ public interface SkillHandler {
 
     void decreaseExperience(SkillDatabaseEntity skillDatabaseEntity, int value);
 
-    long getExperience(SkillDatabaseEntity skillDatabaseEntity);
+    int getExperience(SkillDatabaseEntity skillDatabaseEntity);
 
     SkillType getSupportedSkillType();
 }
